Fix erroneous declaration of i/o error on devices.
errno was being checked to see if there was an error on the input
or output operations. That was indicating an error when reading
from data sourcing devices, e.g. /dev/null, /dev/zero, even though
there was no error. The mistake was not checking ferror() on the
respective streams first, which must be done, and return true,
before consulting errno.